2-create-reusable-components: refactor imports
