--- shared/components/index.ts
@@ -1,6 +1,0 @@
-export * from './CustomLink';
-export * from './Button';
-export * from './Input';
-export * from './ImageUploader';
-export * from './ErrorNotifications';
-export * from './Card';
